requirements.txt
fastapi
uvicorn
jinja2
requests
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
fpdf2
python-dotenv
